Adds note about unimplemented feature.
The feature being used in this code block is not yet implemented in the latest build.  Either it should be removed from the documentation or annotated as such.  I have chosen to annotate it so it will be easy to add the documentation back in later, but removal also works if that is preferred.