I would like to fix unfair comparison and mistakes
I found many mistakes in this page, I would like to fix mistake.

> In Scala, sometimes it’s very hard to tell what’s happening in your code without using a debugger, because too many implicits get into the picture

Because implicit conversions and parameters are resolved in compile time, the debugger is useless.  It's wrong.

> Path-dependent types

It's not a feature. it's a part of type system.

> Existential types
>   Type projections are a very special case

The feature of existential types in Scala is quite limited because it is introduced to only support Java wildcard.  Then, Scala's existential types are not a very special case of Type projections. it is quite similar feature with Type projections.

> Yield operator

Scala has no 'yield' operator(`yield` is just a keyword used in for-expression.  Do you researched about Scala in a serious manner really?

> We plan to support [Project Valhalla](http://openjdk.java.net/projects/valhalla/) once it is released as part of the JDK

In serious comparison to actual language, a word such as `plan` should not be used.  If Scala creator say 'Built-in XML will be planned to removed in future', do you write it in language comparison really?