added appropriate reference to https://kotlinlang.org/docs/tutorials/android-plugin.html
I know the Kotlin Android Extensions plugin is obsolete, as its functionality has now been merged with the main Kotlin plugin. So, I replaced the reference to the former with a ref to the latter and added a link to the page where one can read on how to use the functionality.(Note: this commit builds on my previous changes. See my previous commit for details of what I did in that commit.)