Add constructor example to the `@JvmOverloads` docs
Constructors overloads generation seem to be a common use case for `@JvmOverloads` (though the Github search is complicated by forks). But annotating a primary constructor is not syntactically trivial. To save the readers from searching for the syntax (which was not mentioned or linked here), it seems rational to just add the constructors to the example.