Update comparison-to-java.md
There is this endless discussion about whether or not to add the ternary operator to Kotlin https://discuss.kotlinlang.org/t/ternary-operator/2116/122 and someone pointed out, that the docs don't list this as a difference to java. 
It seems like the ternary operator wont be added so I think it is a good idea to state this here, as it makes changing from java easier.