Update classes.md
The page lists the *Design and document for inheritance or else prohibit it* Effective Java entry as being item 17 in "Effective Java".  This is technically an incorrect reference since listing the book as "Effective Java" implies the first edition, in which the mentioned item is #15. The citation should therefore be "Effective Java, 2nd Edition, Item #17". But given that the third edition was released in this past December, I think it, as the latest edition. should be cited. In the third edition it's item #19. Thus I recommend chaining the cite to  "Effective Java, 3rd Edition, Item #19".