feat(nav.yml): support for includes in the _nav.yml
We need to split the global `_nav.yml` into parts to share site sources between repositories. We assume the site is included under the `pages/` folder, thus there is no need to change the way `.md` files are rendered.

Right now the following is supported in the root _nav.yml:
```
    - title: Native
      description: Kotlin Native and iOS
      '@include': /docs/reference/native/_nav.yml
      '@prefix': /docs/reference/native/

```
The `@include` and `@prefix` elements will be replaced with the `content` element, the included yml file contents are placed under the created `content` element. We assume the imported yml is list.

`.yml` file to import:
```
- url: INTEROP.html
  title: C Interop
- url: OBJC_INTEROP.md
  title: Objective-C/Swift Interop
```

The `@prefix` parameter helps to fix `url` elements of the imported `.yml` to include prefix paths, e.g., `url: foo` will be imported as `url: /docs/reference/native/foo`.