Remove ambiguity from infix function calls predence examples
The examples showing precedence of infix function calls have some ambiguity due to the use of the word and which is a bitwise operation keyword in Kotlin to join two equivalent expressions. The example block is enclosed in indented quoting (>) which drowns the monospace font otherwise differentiating the conjuctive use of the word and between the monospaced expressions. It all appears as though one long expressions and takes you a few minutes to get the gist of what's going on.