Clarification of `Id` when choosing good names.
Nit pick: When choosing good names for things, `ID` is often a point of contention. 
I was happy to see that the coding conventions doc has a specific section for acronyms and follows what I'd expect. I think it bares clarification that `Id` is not an acronym, it's an abbreviation.