Add mention of caveat with default interface methods in delegation page
I got bit by this caveat today, and I was quite surprised to discover
that this behavior is actually by design, and is clearly documented
on the Java interop page -- I had never even considered looking there.

So I suggest adding a small call-out to the delegation page, so that
others may discover it if they run into the same issue.