Update control-flow.md
The when expression replaces or is similar to the switch statement in C-like languages. The problem was with the "the switch operator of C-like languages". *switch* is not an operator, it is a statement. (And it is an expression in Kotlin, Haskell, Elm and the like. Which is actually quite interesting.)