Updated: Compile Kotlin and Java sources
I think this is the modern way to run Kotlin and Java together in the same project.  IntelliJ IDEA warns if you try to do it the old way.  The new is also simpler; you don't even need a maven-compiler-plugin section.

I also added a note to use one of: jar-file-plugin or assembly-plugin or maven-shade-plugin