minor change in introduction to variance
List<? super String> accepts objects of type List of either String type, or supertype of String type, but not a supertype of the object type. But it is written in such a way that it accepts supertype of Object which is fundamentally wrong. So, it should be List<String> instead of List<Object>. That is List<? super String> is a supertype of List<String>.