Clarify fun interfaces vs. type aliases section on which one to use
The current text creates an impression that functional interfaces should be preferred to typealises, since they are more flexible, which is not really the case. Clarification is added to explain in which circumstances the use of functional interfaces is warranted, and in which a plain functional type shall be used.