KT-47068: Add schema for data/user-groups.yml
It helps ensuring that the data is valid to some extent, e.g. helps
avoid typos in field names or missing fields.

After this commit, it's possible to e.g. point IntelliJ IDEA to use
this schema when editing `data/user-groups.yml` and the IDE will show
warnings in case something's wrong. Ultimately, the file should be
validated on each pull request and on server startup, to prevent
consuming corrupted data.