KT-47068: Allow specifying positions for user groups
After this change, it's possible to specify positions for each user
group and it's recommended to do for newly added groups. Backfill for
existing user groups will be done in a separate change.

Names of the new fields were inspired by what's already in
`kotlinconf.yml`, for consistency.

Tricity Kotlin User Group was augmented with position as an example.