Replace dependency on Node Buffer API
This fixes an issue: https://kotlinlang.slack.com/archives/C0B8L3U69/p1628876549226400

The issue is likely cause by the project being created by IDEA, which produces

```
js(LEGACY) {
    browser {
        commonWebpackConfig {
            cssSupport.enabled = true
        }
    }
}
```

in build.gradle.kts. Note that this uses a `browser` instead of `nodejs` target. Because of this, running the tests as described in the document (before this commit) results in failure:

```
ReferenceError: Buffer is not defined
```

This commit fixes that issue by replacing the used Node API for a browser API: `btoa`.