Update collection-aggregate.md
Update collection-aggregate.md to fix the incorrectly documented relationship between "by" max/min functions (`maxByOrNull`) and "of" max/min functions (`maxOfOrNull`). The current documentation says that `maxByOrNull` and `maxOf` are the same except for how they handle empty corrections. I believe this is incorrect because `maxOf` also returns a selector return value, not a collection element. In addition, the current documentation does not mention some functions, such as `maxOfOrNull` and `maxOfWithOrNull`.

This is my first time contributing to the docs; please let me know if I made a mistake or am misunderstanding something.