fix: fix broken link to java.lang.Record documentation (#2650)
* Fix broken link to java.lang.Record documentation

The JVM records tutorial links to the Java documentation for
`java.lang.Record`, but that link no longer works. These updates change
the link to the current JDK 16 documentation.

(This change does not update the link to point to newer JDK versions'
documentation, so that it keeps with the original intention of this
page. JVM records came out in JDK 16.)

* update: update jvm-records.md

Co-authored-by: Andrey Polyakov <koshachy@gmail.com>