Change a block code for better consistancy
Property delegate requirements -> resourceDelegate()

Throughout this part, in code examples, "Resource" class has been used for properties type. the function name is resourceDelegate() too that adds up to confusion. also the function is an example of using anonymous objects instead of classes that has been used in the previous example, so I think it's more readable if the code examples be consistent.