clarify sealed class inheritance restrictions
Indicate the base package restriction around sealed class in the first paragraph.

Sealed classes can only be inherited from within the same package and module in which the sealed class is defined. This implies:

1. that classes within the same compilation unit (module) but not within the base package cannot inherit from them
2. that classes in the same base package but not in the same compilation unit can also not inherit from them.

Previously, this paragraph only indicated the second restriction (2.), that the class needs to be within the same module. This change tries to also capture (1.) in this first paragraph. The same restrictions are explained in greater detail in a later section.