equality.md: add a clarification to the section on structural equality
Because the structural equality operator `==` desugars to `equals`, there are some cases in which its underlying behavior will actually be *referential*. In fact, this is the default for non-data classes.

The name "structural equality", while it makes sense when considering data classes, may lead to confusion for this reason. I've added a note to clarify that, in spite of the name of the operator, there may be cases in which the underlying behavior of the comparison is not strictly structural in nature.