Refactor ‘save’ and ‘load’ methods for valid syntax.
Enhance the "Definitely non-nullable types" section, which includes the following Java code and an example illustrating the most typical use cases for declaring definitely non-nullable types in Kotlin:

public interface Game<T> {
    public T save(T x) {}
    @NotNull
    public T load(@NotNull T x) {}
} 

Java does not permit the implementation of methods in interfaces unless they are static or default, which is why this code does not compile. The public qualifier is also unnecessary because all Java interface methods are implicitly public.